intento 1: jajas

+++ Arkanoid 2.0/HolaSDL/Paddle.h
@@ -16,7 +16,7 @@ private:
 	Game* game = nullptr;
 
 public:
-	Paddle(Vector2D p, int h, int w, Texture* t, Game* g, Vector2D d, int rL, int lL, int s) : MovingObject(p, h, w, t, d, s), game(g), rightLimit(rL), leftLimit(lL) {};
+	Paddle(Vector2D p, int h, int w, Texture* t, Game* g, Vector2D d, int rL, int lL, double s) : MovingObject(p, h, w, t, d, s), game(g), rightLimit(rL), leftLimit(lL) {};
 	virtual void update();
 	virtual void loadFromFile(ifstream& loadFile);
 	virtual void saveToFile(ofstream& saveFile);
